Ported makecatalogs from python to coffeescript.
Now running makecatalogs doesn't have to shell out to python, which means there's no longer a requirement for Munki to be installed on the server...which also means, we should have _true_ linux support now.

Also added get(), set() and key() methods to the MandrillSettings collection object to simplify the rest of the code. The get() method only needs one argument, the key you're trying to retrive, but if you pass it a second argument, that will be returned in the event the key you're looking for isn't found. The set() method will take care of updating or inserting records as needed. All of the methods implicitly retain reactivity.