Hardened GitBroker
Commit messages were previously being escaped before being sent to the
shell, but when I tried some markdown with back ticks, I realized how
large of a mountain there was to climb in order to make that
fool-proof. So instead, we’re using the slower, but much more reliable
method of a commit file.